fix: uint underflow

+++ Player/Logic/Map.cs
@@ -14,8 +14,8 @@ internal static class Map
         bases.OrderBy(x => CalculateDistanceBetweenBases(source, x)).Skip(1).ToList();
 
     internal static uint CalculateMoveCost(Base source, Base destination, PathConfig pathConfig) =>
-        Math.Max(0,
-            (CalculateDistanceBetweenBases(source, destination) - pathConfig.GracePeriod) * pathConfig.DeathRate);
+        (uint) Math.Max(0,
+            (int) (CalculateDistanceBetweenBases(source, destination) - pathConfig.GracePeriod) * pathConfig.DeathRate);
 
     internal static uint CalculateAmountAfterMove(Base source, Base destination, PathConfig pathConfig, uint amount) =>
         amount - CalculateMoveCost(source, destination, pathConfig);
